Added test for color enum, moved files out of 'java' package
