+++ src/main/Enums/Colors.java
@@ -1,0 +1,13 @@
+package Enums;
+
+public enum Colors {
+    BLACK("BLACK"),
+    WHITE("WHITE");
+
+    private final String color;
+
+    Colors(String color){
+        this.color = color;
+    }
+
+}
