+++ src/main/Main.java
@@ -1,5 +1,3 @@
-package main.java;
-
 public class Main {
     public static void main(String args[]){
