+++ src/test/MainTest.java
@@ -1,5 +1,3 @@
-package java;
-
 import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
